pyproject.toml
[tool.poetry]
name = "timed"
version = "3.0.7"
description = "Timetracking software"
repository = "https://github.com/adfinis/timed-backend"
authors = ["Adfinis AG"]
license = "AGPL-3.0"
readme = "README.md"
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Environment :: Console",
	"Intended Audience :: Developers",
	"Intended Audience :: Information Technology",
	"License :: OSI Approved :: GNU Affero General Public License v3",
	"Natural Language :: English",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: 3.8",
]
include = ["CHANGELOG.md"]

[tool.poetry.dependencies]
python = "^3.11"
python-dateutil = "^2.9.0"
django = "^4.2.11"
# might remove this once we find out how the jsonapi extras_require work
django-cors-headers = "^4.3.1"
django-filter = "^24.2"
django-multiselectfield = "^0.1.12"
django-prometheus = "^2.3.1"
djangorestframework = "^3.15.1"
djangorestframework-jsonapi = "^6.1.0"
mozilla-django-oidc = "^4.0.1"
psycopg2-binary = "^2.9.9"
pytz = "^2024.1"
pyexcel-webio = "^0.1.4"
pyexcel-io = "^0.6.6"
django-excel = "^0.0.10"
django-nested-inline = "^0.4.5"
pyexcel-ods3 = "^0.6.1"
pyexcel-xlsx = "^0.6.0"
pyexcel-ezodf = "^0.3.4"
django-environ = "^0.11.2"
django-money = "^3.4.0"
python-redmine = "^2.5.0"
sentry-sdk = "^1.45.0"
whitenoise = "^6.6.0"
django-hurricane = "^1.5.0"
openpyxl = "3.0.10" # TODO: dependency of `pyexcel-xlsx` Remove as soon as https://github.com/pyexcel/pyexcel-xlsx/issues/52 is resolved.

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"
coverage = "7.4.4"
django-extensions = "3.2.3"
factory-boy = "3.3.0"
ipdb = "0.13.13"
pdbpp = "0.10.3"
pytest = "8.1.1"
pytest-cov = "5.0.0"
pytest-django = "4.8.0"
pytest-env = "1.1.3"
# needs to stay at 2.1.0 because of wrong interpretation of parameters with "__"
pytest-factoryboy = "2.1.0"
pytest-freezegun = "0.4.2"
pytest-mock = "3.14.0"
pytest-randomly = "3.15.0"
requests-mock = "1.12.1"
snapshottest = "0.6.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "timed.settings"
addopts = "--reuse-db --randomly-seed=1521188767 --randomly-dont-reorganize"
env = [
    "DJANGO_OIDC_USERNAME_CLAIM=sub"
]
filterwarnings = [
    "error::DeprecationWarning",
    "error::PendingDeprecationWarning",
    "ignore: The USE_L10N setting is deprecated. Starting with Django 5.0, localized formatting of data will always be enabled.:django.utils.deprecation.RemovedInDjango50Warning",
    # following is needed beceause of https://github.com/mozilla/mozilla-django-oidc/pull/371
    "ignore:distutils Version classes are deprecated:DeprecationWarning", # deprecation in pytest-freezegun
    "ignore:.*is deprecated in favour of new moneyed.l10n.format_money.*",
    "ignore:.*invalid escape sequence.*",
    "ignore:pkg_resources is deprecated as an API:DeprecationWarning",
]

[tool.coverage.run]
source = ["."]

[tool.coverage.report]
fail_under = 100
exclude_lines = [
    "pragma: no cover",
    "pragma: todo cover",
    "def __str__",
    "def __unicode__",
    "def __repr__",
]
omit = [
    "*/migrations/*",
    "*/apps.py",
    "*/admin.py",
    "manage.py",
    "timed/redmine/management/commands/import_project_data.py",
    "timed/settings_*.py",
    "timed/wsgi.py",
    "timed/forms.py",
    "setup.py",
]
show_missing = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
